pyproject.toml
[project]
name = "focusassist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq",
    "ollama",
    "Pillow",
    "python-dotenv",
    "pywin32"
]
